requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
psycopg2-binary==2.9.9
werkzeug==3.0.1
pillow==10.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
flask
flask-login
flask-sqlalchemy
gunicorn
pillow
psycopg2-binary
python-dotenv
werkzeug
